[COMM-629] Add SVG cog wheel to moderation actions button
The SVG is taken from `gear-stroke.svg`[1] from the Zendesk Garden SVG
Icons repository. Another possibility is `gear-fill.svg`[2] from the
same repo.

I've changed the markup from a span to a button.  It has multiple
effects, one of which is making it possible to focus it with the
keyboard.  I've also made the button 44px by 44px.  I've also removed
the rules that removed the outline (when it's focused).

TODO:
===

Include `aria-pressed`:

> The aria-pressed attribute values of true or false identify a button
  as a toggle button. A menu button is a button that controls a menu and
  has an aria-haspopup property attribute set to either menu or true.

-- https://developer.mozilla.org/en-US/docs/Web/Accessibility/ARIA/Roles/button_role

[1]: https://github.com/zendeskgarden/svg-icons/blob/master/src/16/gear-stroke.svg
[2]: https://github.com/zendeskgarden/svg-icons/blob/master/src/16/gear-fill.svg