(#1) Add choco-theme to Help Guide
This adds the necessary bits to make the Zendesk Help Guide more
choco-themed.

While the css/js/images have been modified, the base template has
mostly been left unchanged. This was done intentionally to hopefully
allow updates to the base theme to be much more graceful. While we want
this to be choco-themed, we still want to be able to pull in new
features from Zendesk when needed from the source repository.