fix: increase border contrast ratio for more theme elements
Initially I added these changes to Help Center directly. It turns
out that doing so would amount to a breaking change for customers
who assume that these elements will remain the same color in their
custom themes. To err on the side of caution, I am adding them
to the Copenhagen theme instead.