fix(new-request-form): fixed rendering issues with conditional fields
In some cases, where fields get moved around because of end user conditions,
the rendering of the fields was not working as expected. This was due to the
fact that the key was not being passed correctly to the components. This fix ensures
that all components are provided with the correct key to prevent issues with rendering.